added getServerSideProps for api call added next Head component to each page for SEO
